server-side/cleaverwall/requirements.txt
django
djangorestframework

tensorflow
numpy
pefile
# scikt-learn & pandas not needed?
